Remove old ResourceWriters. Remove ResourcePackager, and rename LibraryPackager to ResourcePackager
