+++ Cql/Cql.Packaging/LibraryPackagerFactory.cs
@@ -8,7 +8,7 @@ using Microsoft.Extensions.Options;
 namespace Hl7.Cql.Packaging;
 
 /// <summary>
-/// This creates all services necessary for a <see cref="LibraryPackager"/>.
+/// This creates all services necessary for a <see cref="ResourcePackager"/>.
 /// The idea is not to inject this into service types, it's purpose is to
 /// be one alternative to the .net hosting's <see cref="IServiceProvider"/>.
 /// </summary>
@@ -18,7 +18,7 @@ internal class LibraryPackagerFactory : LibrarySetExpressionBuilderFactory
     private readonly Lazy<CSharpCodeStreamPostProcessor?> _cSharpCodeStreamPostProcessor;
     private readonly Lazy<FhirResourcePostProcessor?> _fhirResourcePostProcessor;
     private readonly Lazy<AssemblyCompiler> _assemblyCompiler;
-    private readonly Lazy<LibraryPackager> _libraryPackager;
+    private readonly Lazy<ResourcePackager> _libraryPackager;
     private readonly Lazy<CqlTypeToFhirTypeMapper> _cqlTypeToFhirTypeMapper;
 
     public LibraryPackagerFactory(
@@ -43,7 +43,7 @@ internal class LibraryPackagerFactory : LibrarySetExpressionBuilderFactory
                 : null);
         _cSharpSourceCodeWriter = Deferred(() => new CSharpLibrarySetToStreamsWriter(Logger<CSharpLibrarySetToStreamsWriter>(), FhirTypeResolver));
         _assemblyCompiler = Deferred(() => new AssemblyCompiler(CSharpLibrarySetToStreamsWriter, TypeManager, CSharpCodeStreamPostProcessor));
-        _libraryPackager = Deferred(() => new LibraryPackager(FhirTypeResolver, AssemblyCompiler, LibrarySetExpressionBuilder, FhirResourcePostProcessor));
+        _libraryPackager = Deferred(() => new ResourcePackager(FhirTypeResolver, AssemblyCompiler, LibrarySetExpressionBuilder, FhirResourcePostProcessor));
 
 
         static Lazy<T> Deferred<T>(Func<T> deferred) => new(deferred);
@@ -62,5 +62,5 @@ internal class LibraryPackagerFactory : LibrarySetExpressionBuilderFactory
 
     public AssemblyCompiler AssemblyCompiler => _assemblyCompiler.Value;
 
-    public LibraryPackager LibraryPackager => _libraryPackager.Value;
+    public ResourcePackager ResourcePackager => _libraryPackager.Value;
 }
