+++ Cql/Cql.Packaging/PublicAPI.Unshipped.txt
@@ -1,6 +1,4 @@
 #nullable enable
-abstract Hl7.Cql.Packaging.ResourceWriters.ResourceWriter.WriteResource(Hl7.Fhir.Model.Resource! resource) -> void
-abstract Hl7.Cql.Packaging.ResourceWriters.ResourceWriter.WriteResources(System.Collections.Generic.IEnumerable<Hl7.Fhir.Model.Resource!>! resources) -> void
 Hl7.Cql.Packaging.AssemblyLoadContextExtensions
 Hl7.Cql.Packaging.CqlTypeToFhirMapping
 Hl7.Cql.Packaging.CqlTypeToFhirMapping.CqlType.get -> Hl7.Cql.Primitives.CqlPrimitiveType?
@@ -28,45 +26,15 @@ Hl7.Cql.Packaging.PostProcessors.FhirResourceWriterOptions.OutDirectory.get -> S
 Hl7.Cql.Packaging.PostProcessors.FhirResourceWriterOptions.OutDirectory.set -> void
 Hl7.Cql.Packaging.PostProcessors.FhirResourceWriterOptions.OverrideDate.get -> System.DateTime?
 Hl7.Cql.Packaging.PostProcessors.FhirResourceWriterOptions.OverrideDate.set -> void
-Hl7.Cql.Packaging.ResourcePackageArgs
-Hl7.Cql.Packaging.ResourcePackageArgs.AfterPackageMutator.get -> System.Action<Hl7.Fhir.Model.Resource!>?
-Hl7.Cql.Packaging.ResourcePackageArgs.AfterPackageMutator.init -> void
-Hl7.Cql.Packaging.ResourcePackageArgs.CqlDir.get -> System.IO.DirectoryInfo!
-Hl7.Cql.Packaging.ResourcePackageArgs.ElmDir.get -> System.IO.DirectoryInfo!
-Hl7.Cql.Packaging.ResourcePackageArgs.ResourceCanonicalRootUrl.get -> string?
-Hl7.Cql.Packaging.ResourcePackageArgs.ResourceCanonicalRootUrl.init -> void
-Hl7.Cql.Packaging.ResourcePackageArgs.ResourcePackageArgs(Hl7.Cql.Packaging.ResourcePackageArgs! original) -> void
-Hl7.Cql.Packaging.ResourcePackageArgs.ResourcePackageArgs(System.IO.DirectoryInfo! elmDir, System.IO.DirectoryInfo! cqlDir, System.Action<Hl7.Fhir.Model.Resource!>? afterPackageMutator = null, string? resourceCanonicalRootUrl = null) -> void
-Hl7.Cql.Packaging.ResourceWriters.CSharpResourceWriter
-Hl7.Cql.Packaging.ResourceWriters.CSharpResourceWriter.CSharpResourceWriter(Microsoft.Extensions.Options.IOptions<Hl7.Cql.CodeGeneration.NET.CSharpCodeWriterOptions!>! options, Microsoft.Extensions.Logging.ILogger<Hl7.Cql.Packaging.ResourceWriters.CSharpResourceWriter!>! logger) -> void
-Hl7.Cql.Packaging.ResourceWriters.CSharpResourceWriter.CSharpResourceWriter(System.IO.DirectoryInfo! outDirectory, Microsoft.Extensions.Logging.ILogger<Hl7.Cql.Packaging.ResourceWriters.CSharpResourceWriter!>! logger) -> void
-Hl7.Cql.Packaging.ResourceWriters.FhirResourceWriter
-Hl7.Cql.Packaging.ResourceWriters.FhirResourceWriter.FhirResourceWriter(Microsoft.Extensions.Options.IOptions<Hl7.Cql.Packaging.PostProcessors.FhirResourceWriterOptions!>! options, Microsoft.Extensions.Logging.ILogger<Hl7.Cql.Packaging.ResourceWriters.FhirResourceWriter!>! logger) -> void
-Hl7.Cql.Packaging.ResourceWriters.FhirResourceWriter.FhirResourceWriter(System.IO.DirectoryInfo! outDirectory, Microsoft.Extensions.Logging.ILogger<Hl7.Cql.Packaging.ResourceWriters.FhirResourceWriter!>! logger) -> void
-Hl7.Cql.Packaging.ResourceWriters.ResourceWriter
-Hl7.Cql.Packaging.ResourceWriters.ResourceWriter.ResourceWriter() -> void
 override Hl7.Cql.Packaging.CqlTypeToFhirMapping.Equals(object? obj) -> bool
 override Hl7.Cql.Packaging.CqlTypeToFhirMapping.GetHashCode() -> int
 override Hl7.Cql.Packaging.CqlTypeToFhirMapping.ToString() -> string!
-override Hl7.Cql.Packaging.ResourcePackageArgs.Equals(object? obj) -> bool
-override Hl7.Cql.Packaging.ResourcePackageArgs.GetHashCode() -> int
-override Hl7.Cql.Packaging.ResourcePackageArgs.ToString() -> string!
-override Hl7.Cql.Packaging.ResourceWriters.CSharpResourceWriter.WriteResource(Hl7.Fhir.Model.Resource! resource) -> void
-override Hl7.Cql.Packaging.ResourceWriters.CSharpResourceWriter.WriteResources(System.Collections.Generic.IEnumerable<Hl7.Fhir.Model.Resource!>! resources) -> void
-override Hl7.Cql.Packaging.ResourceWriters.FhirResourceWriter.WriteResource(Hl7.Fhir.Model.Resource! resource) -> void
-override Hl7.Cql.Packaging.ResourceWriters.FhirResourceWriter.WriteResources(System.Collections.Generic.IEnumerable<Hl7.Fhir.Model.Resource!>! resources) -> void
 static Hl7.Cql.Packaging.AssemblyLoadContextExtensions.Run(this System.Runtime.Loader.AssemblyLoadContext! assemblyContext, string! library, string! version, Hl7.Cql.Runtime.CqlContext! ctx) -> System.Collections.Generic.IDictionary<string!, object?>!
 static Hl7.Cql.Packaging.CqlTypeToFhirMapping.operator !=(Hl7.Cql.Packaging.CqlTypeToFhirMapping? left, Hl7.Cql.Packaging.CqlTypeToFhirMapping? right) -> bool
 static Hl7.Cql.Packaging.CqlTypeToFhirMapping.operator ==(Hl7.Cql.Packaging.CqlTypeToFhirMapping? left, Hl7.Cql.Packaging.CqlTypeToFhirMapping? right) -> bool
 static Hl7.Cql.Packaging.ElmLibraryExtensions.EvaluateExpression(this Hl7.Cql.Elm.Library! library, Hl7.Cql.Elm.Expression! expression, Hl7.Cql.Runtime.CqlContext! context, Microsoft.Extensions.Logging.ILoggerFactory! logFactory) -> object?
 static Hl7.Cql.Packaging.PostProcessors.FhirResourceWriterOptions.BindConfig(Hl7.Cql.Packaging.PostProcessors.FhirResourceWriterOptions! opt, Microsoft.Extensions.Configuration.IConfiguration! config) -> void
-static Hl7.Cql.Packaging.ResourcePackageArgs.operator !=(Hl7.Cql.Packaging.ResourcePackageArgs? left, Hl7.Cql.Packaging.ResourcePackageArgs? right) -> bool
-static Hl7.Cql.Packaging.ResourcePackageArgs.operator ==(Hl7.Cql.Packaging.ResourcePackageArgs? left, Hl7.Cql.Packaging.ResourcePackageArgs? right) -> bool
 virtual Hl7.Cql.Packaging.CqlTypeToFhirMapping.<Clone>$() -> Hl7.Cql.Packaging.CqlTypeToFhirMapping!
 virtual Hl7.Cql.Packaging.CqlTypeToFhirMapping.EqualityContract.get -> System.Type!
 virtual Hl7.Cql.Packaging.CqlTypeToFhirMapping.Equals(Hl7.Cql.Packaging.CqlTypeToFhirMapping? other) -> bool
 virtual Hl7.Cql.Packaging.CqlTypeToFhirMapping.PrintMembers(System.Text.StringBuilder! builder) -> bool
-virtual Hl7.Cql.Packaging.ResourcePackageArgs.<Clone>$() -> Hl7.Cql.Packaging.ResourcePackageArgs!
-virtual Hl7.Cql.Packaging.ResourcePackageArgs.EqualityContract.get -> System.Type!
-virtual Hl7.Cql.Packaging.ResourcePackageArgs.Equals(Hl7.Cql.Packaging.ResourcePackageArgs? other) -> bool
-virtual Hl7.Cql.Packaging.ResourcePackageArgs.PrintMembers(System.Text.StringBuilder! builder) -> bool
